chore: remove breakoutRoom file
