--- src/breakoutRooms.ts
@@ -1,5 +1,0 @@
-const breakoutRooms = new Map<string, string>()
-
-export const getBreakoutRooms = (): Map<string, string> => {
-  return breakoutRooms
-}
